requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
zope.interface==7.2
